requirements.txt
# requirements.txt

--extra-index-url https://download.pytorch.org/whl/cpu

# ========= CORE APP =========
streamlit==1.38.0
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1
regex==2025.9.18
fsspec==2025.9.0
requests==2.32.5
pyyaml==6.0.3

# ========= PDF / DOC LOADING =========
pypdf==4.3.1
pdfminer.six==20240706
python-docx==1.1.2
PyMuPDF

# ========= RETRIEVAL =========
faiss-cpu==1.8.0.post1

# ========= EMBEDDINGS =========
sentence-transformers==2.6.1
transformers==4.41.1
tokenizers==0.19.1
huggingface-hub==0.35.1
# Downgraded for Py3.9 compatibility
scikit-learn==1.6.1
# Downgraded for Py3.9 compatibility (1.16.x needs Py3.10+)
scipy==1.13.1
Pillow==10.4.0

# ========= TORCH FAMILY (CPU) =========
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
sentencepiece==0.2.0

# ========= OPTIONAL OCR =========
pytesseract==0.3.13

# ========= LANGCHAIN =========
langchain-core==0.2.9
langchain-community==0.2.4
langchain==0.2.5

# ========= LLM RUNTIME =========
llama-cpp-python

# ========= EXPORT =========
reportlab



psycopg2-binary
sqlalchemy